Long-term purge memory & more efficient entity serialization

+++ src/commands/purge.ts
@@ -2,15 +2,16 @@ import * as discord from 'discord.js';
 import * as commando from 'discord.js-commando';
 import {CCBot, CCBotCommand} from '../ccbot';
 import {localAdminCheck} from '../utils';
+import {PurgeDatabaseEntity} from '../entities/purge-database';
 
 /**
- * Thanks the RFG developers for an awesome game.
+ * Purges the bot's messages.
  */
 export default class PurgeCommand extends CCBotCommand {
     public constructor(client: CCBot) {
         const opt = {
             name: '-general purge',
-            description: 'Cleans up messes made with the bot. Defaults to 1 minute.',
+            description: 'Deletes messages from the bot. Defaults to 1 minute.',
             group: 'general',
             memberName: 'purge',
             args: [
@@ -26,13 +27,17 @@ export default class PurgeCommand extends CCBotCommand {
     }
     
     public async run(message: commando.CommandMessage, args: {seconds: number}): Promise<discord.Message|discord.Message[]> {
+        // Get the database
+        const database: PurgeDatabaseEntity | undefined = this.client.entities.entities['purge-database-manager'] as (PurgeDatabaseEntity | undefined);
+        if (!database)
+            return await message.say('The purge database doesn\'t exist.');
         if (args.seconds <= 1)
             return await message.say('Too short to be practical.');
         if ((360142280068301 % args.seconds) <= 0)
             return await message.say('You summon an evil space zombie! Shamble shamble. It says that amount of time is way too long.');
         // If they exceed this, they might not actually know that this uses seconds
-        if (args.seconds > (this.client.recentOutputCacheTimeout / 1000))
-            return await message.say('Too long. To prevent massive log scanning, purge uses a cache of sent messages. This lasts ' + (this.client.recentOutputCacheTimeout / 1000) + ' seconds.');
+        if (args.seconds > (database.timeMs / 1000))
+            return await message.say('Too long. To prevent massive log scanning, purge uses a cache of sent messages. This lasts ' + (database.timeMs / 1000) + ' seconds.');
         
         if (!localAdminCheck(message)) {
             return await message.say('You aren\'t authorized to do that.');
@@ -40,18 +45,19 @@ export default class PurgeCommand extends CCBotCommand {
             // Typing information is wrong
             if (!message.channel.bulkDelete)
                 return await message.say('That can\'t be done here. (DM channel?)');
-            // Nuke it. I hope you intended this...
-            const collated: string[] = [];
-            const targetTimestamp = message.createdTimestamp - (args.seconds * 1000);
-            for (const msg of this.client.recentOutputCache) {
-                if (msg.channel.id == message.channel.id) {
-                    if (msg.createdTimestamp > targetTimestamp) {
-                        this.client.entities.killEntity('message-' + msg.id, true);
-                        collated.push(msg.id);
+            const channelData = database.channels[message.channel.id];
+            if (channelData) {
+                // Nuke it. I hope you intended this...
+                const collated: string[] = [];
+                const targetTimestamp = message.createdAt.getTime() - (args.seconds * 1000);
+                for (const msgID of channelData) {
+                    if (discord.SnowflakeUtil.deconstruct(msgID).date.getTime() >= targetTimestamp) {
+                        this.client.entities.killEntity('message-' + msgID, true);
+                        collated.push(msgID);
                     }
                 }
+                await message.channel.bulkDelete(collated);
             }
-            await message.channel.bulkDelete(collated);
             return [];
         }
     }
